Remove latch
